Refactor av komponenter. Fikser null-feil på config for gamle layouts
