+++ src/components/_common/top-container/version-history/VersionHistory.less
@@ -1,0 +1,27 @@
+@import (reference) '../../../../global';
+
+.version-history {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    margin-bottom: 1rem;
+
+    &__toggle-chevron {
+        margin-left: 0.5rem;
+    }
+
+    &__spinner-container {
+        margin-top: 1rem;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+    }
+
+    &__spinner {
+        width: 5rem;
+        height: 5rem;
+        margin-top: 0.5rem;
+    }
+}
