+++ src/components/layouts/LayoutContainer.tsx
@@ -21,13 +21,13 @@ export const LayoutContainer = ({
     children,
     ...divElementProps
 }: Props) => {
-    const { descriptor, path, type, config } = layoutProps;
+    const { descriptor, path, type, config = {} } = layoutProps;
 
     const bem = BEM(type);
     const layoutName = descriptor.split(':')[1];
 
     const commonLayoutStyle = getCommonLayoutStyle(config);
-    const paddingConfig = config?.paddingSides?._selected;
+    const paddingConfig = config.paddingSides?._selected;
 
     const editorProps = pageProps.isDraft
         ? {
